# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus Tech Solutions Client Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── USERS ───────────────────────────────────────────────────────────────────
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"

    post:
      operationId: createClient
      tags: [clients]
      summary: Create/onboard a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientPatch"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── PROJECTS ────────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"

    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status/details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoicePatch"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /invoices/{id}/payment:
    post:
      operationId: recordInvoicePayment
      tags: [invoices]
      summary: Record an advance payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPayment"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── AI PRICING ESTIMATOR ────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send a notification (email/whatsapp for milestone)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── USER ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, dev, client]
        clientId:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Client:
      type: object
      required: [id, companyName, businessType, contactName, contactEmail, services, status, createdAt]
      properties:
        id:
          type: string
        companyName:
          type: string
        businessType:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        services:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, onboarding]
        address:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [companyName, businessType, contactName, contactEmail, services]
      properties:
        companyName:
          type: string
        businessType:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        services:
          type: array
          items:
            type: string
        address:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, onboarding]
    ClientPatch:
      type: object
      properties:
        companyName:
          type: string
        businessType:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        services:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, onboarding]
        address:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]

    # ── PROJECT ───────────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      required: [id, invoiceNumber, clientId, status, items, subtotal, tax, total, dueDate, createdAt]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: ["string", "null"]
        projectId:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        amountPending:
          type: number
        advancePayments:
          type: array
          items:
            $ref: "#/components/schemas/AdvancePayment"
        notes:
          type: ["string", "null"]
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AdvancePayment:
      type: object
      required: [amount, method, date, receiptNumber]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash, upi]
        date:
          type: string
        receiptNumber:
          type: string
    InvoiceLineItem:
      type: object
      required: [description, quantity, rate, amount]
      properties:
        description:
          type: string
        quantity:
          type: number
        rate:
          type: number
        amount:
          type: number
    InvoiceInput:
      type: object
      required: [clientId, items, dueDate]
      properties:
        clientId:
          type: string
        projectId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        tax:
          type: number
        notes:
          type: ["string", "null"]
        dueDate:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
    InvoicePatch:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        notes:
          type: ["string", "null"]
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
    RecordPayment:
      type: object
      required: [amount, method, date]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash, upi]
        date:
          type: string
...
    Notification:
      type: object
      required: [id, userId, type, channel, title, message, read, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [milestone, invoice, project_update, system]
        channel:
          type: string
          enum: [in_app, email, whatsapp]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        entityId:
          type: ["string", "null"]
        createdAt:
          type: string
    NotificationInput:
      type: object
      required: [userId, type, channel, title, message]
      properties:
        userId:
          type: string
        type:
          type: string
          enum: [milestone, invoice, project_update, system]
        channel:
          type: string
          enum: [in_app, email, whatsapp]
        title:
          type: string
        message:
          type: string
        entityId:
          type: ["string", "null"]